+++ package.json
@@ -140,7 +140,7 @@
     "child_process": "^1.0.2",
     "crypto-browserify": "^3.12.0",
     "cypress": "^10.6.0",
-    "dotenv": "^16.0.1",
+    "dotenv": "^16.0.2",
     "electron": "^19.0.5",
     "electron-fetch": "^1.7.4",
     "eslint": "~8.7.0",
